This commit was manufactured by cvs2svn to create tag 'VRJ_RELENG_2_0_1_RELEASE'.
git-svn-id: https://vrjuggler.googlecode.com/svn/tags/VRJ_RELENG_2_0_1_RELEASE@18056 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/buildwin32.py
@@ -270,6 +270,8 @@ def setVars():
    return options
 
 def postProcessOptions(options):
+   os.environ['instprefix'] = options['prefix'].replace('\\', '\\\\')
+
    # Check for Boost 1.32 Visual C++ toolset names.
    match = re.compile(r'vc-(\d)_(\d)').match(options['BOOST_TOOL'])
 
@@ -367,7 +369,7 @@ def generateVersionHeaders():
                self.__genParamFile(output, template)
 
       version_re      = re.compile(r'((\d+)\.(\d+)\.(\d+)-(\d+))\s')
-      branch_re       = re.compile(r'BRANCH\s*=\s*(\w+)')
+      branch_re       = re.compile(r'BRANCH\s*=\s*([\w\d-]+)')
       canon_name_re   = re.compile(r'CANON_NAME\s*=\s*(\S.+)')
       vernum_re       = re.compile(r'@VER_NUMBER@')
       major_vernum_re = re.compile(r'@MAJOR_VER_NUMBER@')
@@ -400,7 +402,7 @@ def generateVersionHeaders():
          for line in params:
             match = self.branch_re.match(line)
             if match is not None:
-               branch_name = match.group(1)
+               branch = match.group(1)
                continue
 
             match = self.canon_name_re.match(line)
@@ -757,12 +759,14 @@ def installDir(startDir, destDir, allowedExts = None, disallowedExts = None,
    os.chdir(cwd)
 
 def installLibs(srcRoot, destdir,
+                buildPlatforms = ['Win32', 'x64'],
                 buildTypes = ['ReleaseDLL', 'DebugDLL', 'Release', 'Debug'],
                 extensions = ['.dll', '.lib']):
-   for t in buildTypes:
-      srcdir = os.path.join(srcRoot, t)
-      if os.path.exists(srcdir):
-         installDir(srcdir, destdir, extensions)
+   for p in buildPlatforms:
+      for t in buildTypes:
+         srcdir = os.path.join(srcRoot, p, t)
+         if os.path.exists(srcdir):
+            installDir(srcdir, destdir, extensions)
 
 def installExternal(prefix):
    # Install Doozer (even though it probably won't be used).
@@ -830,6 +834,12 @@ def installTweek(prefix):
       shutil.copy2(os.path.join(srcroot, f), destdir)
 
 def installTweekJava(prefix):
+   # Create an empty bin\beans directory in prefix. The Tweek Java GUI still
+   # searches this directory for backwards compatibility with much older
+   # versions of the code. This is not required, but it does eliminate a
+   # Tweek Java GUI startup warning.
+   mkinstalldirs(os.path.join(prefix, 'bin', 'beans'))
+
    srcdir = os.path.join(gJugglerDir, 'vc7', 'Tweek_Java')
 
    if os.path.exists(os.path.join(srcdir, 'Tweek.jar')):
@@ -858,11 +868,14 @@ def installTweekJava(prefix):
          shutil.copy2(os.path.join(srcdir, j), destdir)
 
       # Install the tweek_jni DLL.
-      dll = os.path.join(srcdir, 'tweek_jni', 'ReleaseDLL', 'tweek_jni.dll')
-      arch = os.environ['PROCESSOR_ARCHITECTURE']
-      destdir = os.path.join(destdir, arch)
-      mkinstalldirs(destdir)
-      shutil.copy2(dll, destdir)
+      for p in ['Win32', 'x64']:
+         dll = os.path.join(srcdir, 'tweek_jni', p, 'ReleaseDLL',
+                            'tweek_jni.dll')
+         if os.path.exists(dll):
+            arch = os.environ['PROCESSOR_ARCHITECTURE']
+            destdir = os.path.join(destdir, arch)
+            mkinstalldirs(destdir)
+            shutil.copy2(dll, destdir)
 
       destdir = os.path.join(prefix, 'share', 'tweek', 'beans')
       mkinstalldirs(destdir)
@@ -1068,12 +1081,16 @@ def installSonixPlugins(prefix):
    destdir_opt = os.path.join(prefix, 'lib', 'snx', 'plugins', 'opt')
 
    srcroot = os.path.join(gJugglerDir, 'vc7', 'Sonix', 'OpenAL')
-   installLibs(srcroot, destdir_dbg, ['DebugDLL'], ['.dll'])
-   installLibs(srcroot, destdir_opt, ['ReleaseDLL'], ['.dll'])
+   installLibs(srcroot, destdir_dbg, buildTypes = ['DebugDLL'],
+               extensions = ['.dll'])
+   installLibs(srcroot, destdir_opt, buildTypes = ['ReleaseDLL'],
+               extensions = ['.dll'])
 
    srcroot = os.path.join(gJugglerDir, 'vc7', 'Sonix', 'Audiere')
-   installLibs(srcroot, destdir_dbg, ['DebugDLL'], ['.dll'])
-   installLibs(srcroot, destdir_opt, ['ReleaseDLL'], ['.dll'])
+   installLibs(srcroot, destdir_dbg, buildTypes = ['DebugDLL'],
+               extensions = ['.dll'])
+   installLibs(srcroot, destdir_opt, buildTypes = ['ReleaseDLL'],
+               extensions = ['.dll'])
 
 def installGadgeteer(prefix):
    printStatus("Installing Gadgeteer headers, libraries, and samples ...")
@@ -1159,6 +1176,10 @@ def installVRJuggler(prefix):
    installDir(srcdir, destdir, ['.h'])
 
    srcdir  = os.path.join(gJugglerDir, 'vc7', 'VRJuggler', 'vrj')
+   installDir(srcdir, destdir, ['.h'])
+
+   destdir = os.path.join(prefix, 'include', 'deprecated')
+   srcdir  = os.path.join(gJugglerDir, 'modules', 'vrjuggler', 'deprecated')
    installDir(srcdir, destdir, ['.h'])
 
    destdir = os.path.join(prefix, 'lib')
@@ -1329,10 +1350,13 @@ def installMsvcRT(prefix):
 
       # Get *every* MSVC runtime DLL.  This list could be shortened at some
       # point if anyone cares to try.
-      dlls = glob.glob(os.path.join(srcroot, 'System32', 'msvc*.dll'))
+      sys_dir = os.path.join(srcroot, 'System32')
+      dlls = glob.glob(os.path.join(sys_dir, 'msvc*.dll'))
 
       for d in dlls:
          shutil.copy2(d, destdir)
+
+      shutil.copy2(os.path.join(sys_dir, 'dbghelp.dll'), destdir)
    except KeyError, ex:
       printStatus("WARNING: Could not install MSVC runtime DLLs")
       print ex
@@ -1876,7 +1900,7 @@ class GuiFrontEnd:
       devenv_cmd_no_exe = '"%s"' % (devenv_cmd_no_exe)
 
       solution_file = r'"%s"' % os.path.join(gJugglerDir, 'vc7', 'Juggler.sln')
-      build_args = r'/build Debug'
+      build_args = r'/build DebugDLL'
 
       if self.mRoot.CommandFrame.OpenVSCheck.Variable.get() == "No":
          cmd = devenv_cmd_no_exe + ' ' + solution_file + ' ' + build_args
